Change scale from int32_t to double for pango

+++ include/pango.h
@@ -18,10 +18,10 @@
  */
 int escape_markup_text(const char *src, char *dest, int dest_length);
 PangoLayout *get_pango_layout(cairo_t *cairo, const char *font,
-		const char *text, int32_t scale, bool markup);
+		const char *text, double scale, bool markup);
 void get_text_size(cairo_t *cairo, const char *font, int *width, int *height,
-		int32_t scale, bool markup, const char *fmt, ...);
+		double scale, bool markup, const char *fmt, ...);
 void pango_printf(cairo_t *cairo, const char *font,
-		int32_t scale, bool markup, const char *fmt, ...);
+		double scale, bool markup, const char *fmt, ...);
 
 #endif
